+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Make SSLRootCertPath optional
+
 ### Security
 
 ## [v0.16.0] - 2021-03-03